+++ public/components/Guide.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Switch, Route } from 'react-router-dom';
 import {
     Container, Row, Cell, MediaPoints,
-} from "../../src";
+} from "../../index";
 import MainMenu from "./MainMenu";
 import SideMenu from "./SideMenu";
 import "../css/guide.less";
@@ -29,6 +29,10 @@ import GuideInput from "./guide/Input";
 import GuideInputFile from "./guide/InputFile";
 import GuideSelectColor from "./guide/SelectColor";
 import GuideSelectIcon from "./guide/SelectIcon";
+import GuideSelect from "./guide/Select";
+import GuideSwitch from "./guide/Switch";
+import GuideCheckbox from "./guide/Checkbox";
+import GuideRadio from "./guide/Radio";
 
 export default class Guide extends React.Component {
     constructor(props){
@@ -97,6 +101,10 @@ export default class Guide extends React.Component {
 
                                 <Route path='/guide/input' component={GuideInput}/>
                                 <Route path='/guide/input-file' component={GuideInputFile}/>
+                                <Route path='/guide/checkbox' component={GuideCheckbox}/>
+                                <Route path='/guide/switch' component={GuideSwitch}/>
+                                <Route path='/guide/radio' component={GuideRadio}/>
+                                <Route path='/guide/select' component={GuideSelect}/>
                                 <Route path='/guide/select-color' component={GuideSelectColor}/>
                                 <Route path='/guide/select-icon' component={GuideSelectIcon}/>