+++ public/components/guide/InfoButton.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {Adsense, Cell, InfoButton, Table} from "../../../src";
+import {Adsense, Cell, InfoButton, Table} from "../../../index";
 import PrismCode from "../PrismCode";
 import Defs, {tablePropsHeader} from "../Defs";
 import Article from "../Article";