+++ src/components/input/input.jsx
@@ -2,6 +2,12 @@ import React from "react";
 import "./input.less";
 import Button from "../button/button.jsx";
 
+const AutocompleteListItem = ({start, coincidence, end}) => {
+    return (
+        <span>{start}<strong>{coincidence}</strong>{end}</span>
+    )
+};
+
 export default class Input extends React.Component {
     static defaultProps = {
         fieldState: "normal",
@@ -213,10 +219,6 @@ export default class Input extends React.Component {
             <React.Fragment>
                 <div className={'input ' + (focus ? 'focused' : '') + (fieldState === 'error' ? ' invalid ' : fieldState === 'success' ? ' success ' : '') + ' ' + cls}>
 
-                    {prepend !== "" && (
-                        <span className={'prepend ' + clsPrepend}>{prepend}</span>
-                    )}
-
                     <input className={className} {...props} type={inputType} value={value} onChange={this.onChange} ref={ref => this.input = ref} onKeyUp={this.onKeyUp}/>
 
                     {buttons && (
@@ -246,7 +248,11 @@ export default class Input extends React.Component {
                         </div>
                     )}
 
-                    {append !== "" && (
+                    {prepend && (
+                        <span className={'prepend ' + clsPrepend}>{prepend}</span>
+                    )}
+
+                    {append && (
                         <span className={'append ' + clsAppend}>{append}</span>
                     )}
 
@@ -255,15 +261,18 @@ export default class Input extends React.Component {
                             {
                                 this.autocomplete.map(function(item, index) {
                                     const searchIndex = item.toLowerCase().indexOf(value.toLowerCase());
-                                    const itemValue = `${item.substr(0, searchIndex)}<strong>${item.substr(searchIndex, value.length)}</strong>${item.substr(searchIndex + value.length)}`;
 
                                     return (
                                         <div data-item={item}
                                              className={'item ' + clsAutocompleteItem}
                                              key={index}
                                              hidden={item === value || value === '' || searchIndex === -1}
-                                             dangerouslySetInnerHTML={{__html: itemValue}}
-                                             onClick={autocompleteItemClick}/>
+                                             onClick={autocompleteItemClick}>
+                                            <AutocompleteListItem
+                                                start={item.substr(0, searchIndex)}
+                                                coincidence={item.substr(searchIndex, value.length)}
+                                                end={item.substr(searchIndex + value.length)}/>
+                                        </div>
                                     )
                                 })
                             }